Simplified linters file structure 🇲🇨
