+++ src/linters/eslint/__tests__/index.spec.ts
@@ -4,7 +4,7 @@ import { ESLint, loadESLint } from 'eslint'
 
 import { colourLog } from '@Utils/colour-log'
 
-import { lintFiles } from '../lint-files'
+import { lintFiles } from '..'
 
 jest.mock('eslint')
